# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Try our project on this site.",
    "version": "1.0.0",
    "title": "Hotelbuchung"
  },
  "host": "localhost:8007",
  "basePath": "/",
  "tags": [
    {
      "name": "booking",
      "description": "All bookings"
    },
    {
      "name": "city",
      "description": "All cities"
    },
    {
      "name": "hotels",
      "description": "All hotels"
    },
    {
      "name": "rooms",
      "description": "All rooms"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/booking/{bookingnr}": {
      "get": {
        "tags": [
          "booking"
        ],
        "summary": "Get specific bookingnr in bookings",
        "description": "Get specific booking",
        "operationId": "getBookingnr",
        "parameters": [
          {
            "name": "bookingnr",
            "in": "path",
            "description": "bookingnr of booking",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      }
    },
    "/hotels/cities": {
      "get": {
        "tags": [
          "city"
        ],
        "summary": "Get all cities",
        "description": "Get all cities",
        "operationId": "getcities",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      }
    },
    "/hotels/{city}": {
      "get": {
        "tags": [
          "hotels"
        ],
        "summary": "Get all hotels in city",
        "description": "Get all hotels in city",
        "operationId": "gethotelsincity",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "city name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      }
    },
    "/hotels/{city}/{hotelid}": {
      "get": {
        "tags": [
          "hotels"
        ],
        "summary": "Get a specific hotel in specific city",
        "description": "Get a specific hotel in specific city",
        "operationId": "getSpecificHotelinCity",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "city name",
            "required": true,
            "type": "string"
          },
          {
            "name": "hotelid",
            "in": "path",
            "description": "id of hotel in city to return its rooms",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      }
    },
    "/hotels/{city}/{hotelid}/rooms": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Get all rooms for specific hotel",
        "description": "Get all rooms for specific hotel",
        "operationId": "getRoomsOfhotel",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "city name",
            "required": true,
            "type": "string"
          },
          {
            "name": "hotelid",
            "in": "path",
            "description": "id of hotel to return his rooms",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      }
    },
    "/hotels/{city}/{hotelid}/rooms/{roomid}": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Get all rooms for specific hotel",
        "description": "Get a specific rooms for specific hotel",
        "operationId": "getSpecificRoomsOfHotel",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "city name",
            "required": true,
            "type": "string"
          },
          {
            "name": "hotelid",
            "in": "path",
            "description": "id of hotel to return his rooms",
            "required": true,
            "type": "string"
          },
          {
            "name": "roomid",
            "in": "path",
            "description": "id of room to return his properties",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      },
      "put": {
        "tags": [
          "rooms"
        ],
        "summary": "Update rooms guest and status",
        "description": "Update only guest and status",
        "operationId": "updateRoomsGuestAndStatus",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "description": "city name",
            "required": true,
            "type": "string"
          },
          {
            "name": "hotelid",
            "in": "path",
            "description": "id of hotel with rooms to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "roomid",
            "in": "path",
            "description": "id of rooms to update Status and Guest",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "rooms",
            "required": true,
            "schema": {
              "$ref": "#/definitions/rooms"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Error"
          }
        }
      }
    }
  },
  "definitions": {
    "hotels": {
      "type": "object",
      "properties": {
        "hotelid": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "example": "Anaroc Hotels"
        },
        "location": {
          "type": "string",
          "example": "Berlin"
        },
        "rooms": {
          "$ref": "#/definitions/rooms"
        }
      },
      "required": [
        "hotelid",
        "name",
        "location",
        "rooms"
      ],
      "xml": {
        "name": "hotels"
      }
    },
    "rooms": {
      "type": "object",
      "properties": {
        "roomid": {
          "type": "string",
          "example": "100"
        },
        "roomtype": {
          "type": "string",
          "example": "single"
        },
        "price": {
          "type": "number",
          "example": 80
        },
        "status": {
          "type": "string",
          "enum": [
            "free",
            "occupied",
            "locked"
          ],
          "example": "free"
        },
        "guest": {
          "type": "string",
          "example": ""
        },
        "bookedfrom": {
          "type": "string",
          "example": ""
        },
        "bookeduntil": {
          "type": "string",
          "example": ""
        },
        "bookingnr": {
          "type": "string",
          "example": ""
        }
      },
      "required": [
        "roomid",
        "roomtype",
        "price",
        "status"
      ],
      "xml": {
        "name": "rooms"
      }
    }
  },
  "externalDocs": {
    "description": "GitHub Repository of ivp-g7-implementation",
    "url": "https://github.com/strumswell/ivp-g7-implementation"
  }
}